requirements.txt
bcrypt==3.2.0
cached-property==1.5.2
cffi==1.14.3
click==7.1.2
coloredlogs==14.0
cryptography==3.1.1
fabric2==2.5.0
humanfriendly==8.2
invoke==1.4.1
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.3.1
six==1.15.0
